Introduce multiple variations of MavenPublishJavaIntegTest
- only 'main' feature without docs
- only 'main' feature with docs
- multiple features without docs
- multiple features with docs
